API 2

+++ src/components/Homeworks/Session9/Products.vue
@@ -2,6 +2,10 @@
   <div class="container">
     <div class="add-product">
       <el-button type="primary" @click="createProduct()">Thêm mới</el-button>
+      <br>
+      <input v-model="search" class="search" type="text" placeholder="Tìm kiếm sản phẩm" @keydown.enter="searchProduct">
+      <input type="file" accept="image/*" @change="changeFile">
+      <button @click="uploadFile">Submit</button>
     </div>
     <el-dialog
         title="Tạo mới sản phẩm"
@@ -26,7 +30,7 @@
         </span>
     </el-dialog>
     <el-table
-        :data="tableData.filter(data => !search || data.name.toLowerCase().includes(search.toLowerCase()))"
+        :data="tableData"
         style="width: 100%">
       <el-table-column
           label="Tên sản phẩm"
@@ -36,20 +40,24 @@
       <el-table-column
           label="Giá(VNĐ)"
           prop="price">
+        <template slot-scope="scope">
+          {{ Number( scope.row.price).toLocaleString('vi-VN') }} VNĐ
+        </template>
       </el-table-column>
       <el-table-column
           label="Trạng thái"
           prop="description">
       </el-table-column>
       <el-table-column
-          align="right">
-        <template slot="header" slot-scope="scope">
-          <el-input
-              @click="handleEdit(scope.$index, scope.row)"
-              v-model="search"
-              size="mini"
-              placeholder="Type to search"/>
+          label="Ngày tạo"
+          prop="created_at">
+        <template slot-scope="scope">
+          <i class="el-icon-time"></i>
+          {{ formatDate(scope.row.created_at) }}
         </template>
+      </el-table-column>
+      <el-table-column
+          align="right">
         <template slot-scope="scope">
           <el-button
               size="mini"
@@ -57,7 +65,7 @@
           <el-button
               size="mini"
               type="danger"
-              @click="handleDelete(scope.$index, scope.row)">Delete</el-button>
+              @click="handleDelete(scope.$index, scope.row)">Xóa</el-button>
           <el-dialog
               title="Chỉnh sửa sản phẩm"
               :visible.sync="editModal"
@@ -97,6 +105,7 @@
 
 <script>
 import axios from 'axios'
+import moment from 'moment'
 export default {
   data() {
     return {
@@ -109,6 +118,7 @@ export default {
         description: '',
         price: ''
       },
+      image: '',
       total: 0,
       idEdit: '',
       rules: {
@@ -145,7 +155,10 @@ export default {
         if (valid) {
           this.updateProduct(this.idEdit)
           this.editModal = false
-          alert('cập nhật thành công!');
+          this.$message({
+            message: 'cập nhật thành công!',
+            type: 'success'
+          });
         } else {
           console.log('error submit!!');
           return false;
@@ -161,6 +174,7 @@ export default {
           name: this.ruleForm.name,
           description: this.ruleForm.description,
           price: this.ruleForm.price,
+
         }
       }).then(() => {
         this.getData()
@@ -169,13 +183,25 @@ export default {
       });
     },
     handleDelete(index, row) {
-      axios({
-        method: 'delete',
-        url: 'http://vuecourse.zent.edu.vn/api/products/' + row.id,
+      this.$confirm('Bạn có chắc chắn muốn xóa không?', 'Cảnh báo', {
+        confirmButtonText: 'OK',
+        cancelButtonText: 'Cancel',
+        type: 'warning'
       }).then(() => {
-        this.getData()
-      }).catch((error) => {
-        console.log(error);
+        axios({
+          method: 'delete',
+          url: 'http://vuecourse.zent.edu.vn/api/products/' + row.id,
+        }).then(() => {
+          this.getData()
+        }).catch((error) => {
+          console.log(error);
+        });
+        this.$message({
+          type: 'success',
+          message: 'Xóa thành công!'
+        });
+      }).catch(() => {
+
       });
     },
     submitForm(formName) {
@@ -183,7 +209,10 @@ export default {
         if (valid) {
           this.storeProduct()
           this.dialogVisible = false
-          alert('Tạo mới thành công!');
+          this.$message({
+            message: 'Tạo mới thành công!',
+            type: 'success'
+          });
         } else {
           console.log('error submit!!');
           return false;
@@ -230,6 +259,42 @@ export default {
       }).catch((error) => {
         console.log(error);
       });
+    },
+    searchProduct() {
+      axios({
+        method: 'get',
+        url: 'http://vuecourse.zent.edu.vn/api/products?page=1',
+        params: {
+          q: this.search
+        }
+      }).then((response) => {
+        this.tableData = response.data.data.data;
+      }).catch((error) => {
+        console.log(error);
+      });
+    },
+    formatDate (dateString) {
+      return moment(dateString).format('hh:mm | DD/MM/YYYY')
+    },
+    changeFile(e) {
+      if (e.target.files.length) {
+        this.image = e.target.files[0]
+      }
+    },
+    uploadFile() {
+      const frmData = new FormData()
+      frmData.append('name', 'Thành')
+      frmData.append('price', 1)
+      frmData.append('image', this.image)
+      axios({
+        method: 'post',
+        url: 'http://vuecourse.zent.edu.vn/api/products',
+        data: frmData
+      }).then(() => {
+        console.log('success')
+      }).catch((error) => {
+        console.log(error);
+      });
     }
   },
   mounted() {
@@ -241,6 +306,11 @@ export default {
   .container {
     .add-product {
       text-align: left;
+      .search {
+        width: 300px;
+        height: 30px;
+        margin-top: 10px;
+      }
     }
     .pagination {
       text-align: right;
